Master to staging - 12/12/2019 - 2

See merge request team-leaders/aahung-aagahi!259

+++ aahung-mobile/app/src/main/java/com/ihsinformatics/aahung/views/ScoreWidget.java
@@ -80,7 +80,7 @@ public class ScoreWidget extends Widget implements ScoreContract.ScoreViewer {
     @Override
     public WidgetData getValue() {
         Integer scoreValue = Integer.valueOf(binding.score.getText().toString());
-        Integer percentage = Integer.valueOf(binding.percentage.getText().toString());
+        Double percentage = Double.valueOf(binding.percentage.getText().toString());
         Score score = new Score(scoreKey, scoreValue, percentageKey, percentage);
         return new WidgetData(key, score);
     }
